Move template file option to correct subcommands.
For the export subcommand the option was renamed to --output-file because this
makes more sense and it makes handling the differences between input files and
output files easier.

After collecting the input the terminal is reopen as stdin as stdin might be
exhausted but also needed for further queries to the user.