Turn config.Config into a normal class
The singleton interface is removed to make the class and usage less confusing.
The singleton instance is replaced by a global variable in khard/khard.py.
That should make it clear that modifications to this object are persistent (in
contrast to modification to an object that was apearentrly generated on the
fly and not saved: `Config().set_editor(string)`).  It also makes the point
where the object is actually initialized much clearer.