Add a --config option to use an alternate config file
The argparse patch is also removed.

The command line is parsed in two steps in order to make it possible to

- use the config file given on the command line
- display the help message before parsing the config file
- set the debug option before parsing the config file
- get the default command from the config file if none was given on the
  command line

Currently there is one undesired side effect:  If no subcommand is given on
the command line no options for the subcommand can be given on the command
line as well.  Previously this was possible.