Parse the command line remainder instead of reparsing
On the second parsing step the remainder of the first parsing step is parsed
instead of reparsing the whole command line.  This makes it possible to inject
the default command from the config file with minimal effort if no subcommand
was given on the command line.

This fixes the regression mentioned in b2c8273: It is now again possible to
specify arguments to the default command without specifying the default
command itself (with the exception of all options that are valid to the main
parser).

This also makes it possible to specify general options (like --debug or
--config) after the subcommand or its arguments.  They will be processed by
the first parsing run.