Fix help option of subcommands
The remainder of the command line is explicitly collected in the first parsing
step.  This allows the parse_args() method to be used instead of
parse_known_args().  The latter did skip any unknown args in order to process
more elements of sys.argv.  Thereby it did find the help options of
subcommands.  With the remainder argument in the first parser parse_args() can
be forced to stop processing at the first unknown argument.

This has the side effect that no global options can be given after the first
non global argument.  This is the case with either an explicit or
implicit/default subcommand.  So `khard some-search-terms -h` will give the
help of the default subcommand.