Store contacts in a dict instead of a list
Nearly all of the operations that where performed on the old contacts
list can easily be performed on the dict as well. The only exception are
duplicate UIDs:

This changes the handling of contacts with identical UIDs. Before they
could coexist in the same address book. Now only the first one will
ever be available. A warning will be displayed for all subsequent ones.
This also applies to contacts across individual vdirs as the collection
of all address books is implemented as just another subclass of the base
address book.

This is a breaking change for everybody who has contacts with identical
UIDs but the reasoning here is that UIDs should really be unique (at
least across all contacts of one khard installation).

The test show that also empty UIDs are now invalid (will not appear in
the address book).  The tests are adjusted by adding another card and
checking the logging message about the missing contact.