Split address book class
The class is split into an abstract base class and the implementation backed
by a directory of vCards.  The aim is to provide a class hierarchy that can be
extended more easily.  It should be possible to back an AddressBook subclass
with a single vCard file containing multible entries or a database for
example.