Move settings for loading to __init__
This moves several settings that are relevant for loading an address
book from the signature of the load method to the constructor.  This
makes it possible to clean up and simplify the loading of the address
book.  The long term goal could be to only load vcards from disc on
demand during search().

Here the shifting of responsibility is possible because the settings for
loading are only needed once and also will not change during one run of
khard.