Test the warning for missing UID
The warning messages in the tests are currently hard coded and given in
full.  This is very rigorous (so maybe good for the tests) but at the
same time breaks as soon as the logging format changes slightly (so
maybe bad for the tests as we might not care about the *exact* format).