Add high level test for "move"
The test uses a tempfile in order to set up a test fixture with address
books and config.  These are all removed after each test.

This is currently the simplest way I could find to create such a test
environment but maybe unittest.mock can be used more intelligently to
create a fixture that doesn't involve the file system at all.