Simplify derived search regex: remove surrounding ".*"
The regex are later used with re.search() which looks for a match
anywhere in the string anyways.  re.search() is *much* slower with the
old search terms, presumably because it tries to match the regex at any
position in the string and that specific regex again tries to match the
actual search terms (as given on the command line) after any initial
string (matched with the ".*").