Find queries to load address books correctly
It might be that an address book is listed in the target address book
and/or/but not in the source address books.  Again depending on the
presence of source or target search terms this changes the needed
queries to load that address book.

This commit selects the minimal set of queries needed to correctly load
each address book in any of these cases.