Separate source and target queries when loading abooks
The command line arguments might specify different search queries and
address books for source and target contacts (for a merge command for
example).  An address book might be part of either, both or neither.

An example could be the command `khard merge -f -a one,two -A two,three
foo -t bar`, where address book "one" should be loaded with query "foo",
address book "two" should be loaded with "(foo|bar)" and "three" only
with "bar".

The queries used to load the different address books do now contain all
but only those queries that are relevant for this address book.