Add tests for empty fields in yaml input
These are actually regression tests for a bug introduced in
3c39abace194a5a17ce5714127828a2aada16346 and fixed in
73aad4f2f849d350086cd6726695c4ee30907ae7.

Editing a contact did obviously populate the yaml template with empty
fields where there were no corresponding fields in the vcard file.  But
due to the mentioned bug some of these fields where saved back to the
vcard file (as empty vcard values).