Use custom query format instead of regex internally
The search queries from the command line where formally converted to
regex in order to search for contacts.  Because no special regex
features were used (all regex special chars from the command line were
escaped) and in order to prepare for a custom query syntax (see #131)
this implementation is replaced by nested lists of strings.

Strings are matched against the contact with `in`.  The different
individual search terms from the command line are collected into a list
which is then matched recursive with `all()`.  If several different
queries are used to load one address book (for example for a merge
command) `any()` is used to load all possibly needed contacts.