Add parsable argument to Formatter to hide nicknames
The nicknames only shown if the nicknames argument is true and at the
same time the parsable argument is false.

The reasoning here is that the parsable option is mostly used for
integration into other programs (like email clients) and the nickname
should probably not be show in the address field of an email.