Offer name parts when creating contact from emails
When adding contacts through the add-email functionality khard gives the
name of the contact and then asks to create it if missing with the given
email address.

It would be easier to add the contacts if we at least guess the name
parts and prepopulate the fields for the user.