Expand env vars for addressbooks path config
Expand environment variables for the `addressbooks.<name>.path` config
values.  Arbitrarily choose to do environment variable expansion before
tilde expansion.

This may be backwards incompatible for users that use literal dollar
characters (`$`) in their paths.  If the environment variable expansion
is possible, it will be unintentionally expanded.  If it is not
possible, however, it will not be expanded.  So the compatibility issue
is only when the path is expandable (e.g.
`~/classes/english/$LANG/contacts`).  With this change, I do not know of
a way to express these kinds of paths (escaping doesn't work).  Yet, I
doubt anyone would come across this issue, let alone have a literal
dollar sign in their path.

BREAKING CHANGE: parse env vars for addressbooks path config