Replace test for #249 with query tests about spaces
Issue 249 was about queries that would match a part of a contact after
spaces were removed from the contact.  This is generally not desired, as
it touch the topic of fuzzy matching.

If any kind of fuzzy matching or guessing will be implemented it has to
be in the query classes.