Properly split $EDITOR env variable
Before this was taken as a string and passed to `subprocess.Popen`. This
worked when `$EDITOR` was just a command with no arguments appended.
When arguments are appended `subprocess.Popen` requires them to be
passed as a list. This commit introduces the logic for splitting the
`$EDITOR` string into a list as recommended by `subprocess.Popen`'s
documentation.