Specify version (ranges) for dependencies
Atomicwrites and vobject are more or less abandoned and did not have a
release in a long time.  They are pinned to exact versions in case some
change upstream results in an unexpected release.

For the other packages I took a guess what a good version constraint
could be.  They are not thoroughly tested and I did not study the
release notes to check which version we need.