nix: rewrite default derivation and add some check outputs
The main khard derivation is rewritten to not depend on the derivation
in nixpkgs.  This makes it easier to change stuff in the build system or
modify dependencies before a new version is released and packaged in
nixpkgs.

The dependencies are now read from the pyproject.toml file during the
nix evaluation.  This removes the need to specify our dependencies
twice.

Currently only checks with python 3.11 and 3.12 have been added.  Our
locked version of nixpkgs provides more versions but the tests fail in
the nix sandbox for these python versions.  And was not able to switch
the nix build to use the pytestCheckHook because many tests fail with
this.