Replace sys.exit calls with exceptions and return values
Nearly all calls to sys.exit are replaced by khard's custom Cancelled
exception or an explicit return value of type str|int|None.  These are
the values that sys.exit can digest and they are handed outwards to the
exit call with the return value of main().

This should make the code more testable.

Additionally all custom exception are moved to module.