Refactor _get_multi_property into get_all; make name getters nullable
The _list attribute on the underlying vcard object should be faster than
iterating through all attributes ourself.

With this it should be possible to detect if an attribute is present on
the vcard or not.  If it is not present the @property on the wrapper
returns None or an empty list or an empty dict, depending on the
attribute.